script.classes module
